Bug fixes for Linux

+++ src/application/JpGuiController.java
@@ -975,10 +975,10 @@ public class JpGuiController implements Initializable, RefreshScene {
 				os.contains("nix") == true || 
 				os.contains("aix") == true || 
 				os.contains("sunos") == true) {
-			fn += "linux_.sh";
+			fn += "_linux.sh";
 			isWin = false;
 		} else if (os.contains("mac") == true) {
-			fn += "mac_.sh";
+			fn += "_mac.sh";
 			isWin = false;
 		}
 		
@@ -1235,28 +1235,31 @@ public class JpGuiController implements Initializable, RefreshScene {
 				}
 	    	}
 	    	
-	    	prerun = new File(System.getProperty("user.home") + File.separator + "JpGui" + File.separator +
-					"projects" + File.separator + prjName + File.separator + os + "_prerun.sh");
-	    	
-	    	if (prerun.exists() == false) {
-	    		try {
-					FileWriter w = new FileWriter(prerun.getAbsolutePath(), false);
-					w.write("#! /usr/bin/sh\n# PreRun script for project " + prjName + "\n\n");
-					w.write("# Place your code below this line.\n");
-					w.write("# Keep your code above this line.\n\n");
-					w.write("exit 0\n\n");
-					
-					w.write("# Function to copy to input directory (jpackage --input).  Usage: CP filepath\n");
-					w.write("CP()\n{\n");
-					w.write("cp $1 " + jg.currPrj.getString("Application Image Options", jg.platform + " Input") + " >/dev/null 2>&1\n");
-					w.write("if [ $? -ne 0 ]; then\n");
-					w.write("    echo PreRun: Copy of $1 failed.\n");
-					w.write("    exit 1\n");
-					w.write("fi\n}\n");
-					w.close();
-				} catch (IOException e) {
-					e.printStackTrace();
-				}
+	    	if (isWin == false) {
+		    	prerun = new File(System.getProperty("user.home") + File.separator + "JpGui" + File.separator +
+						"projects" + File.separator + prjName + File.separator + os + "_prerun.sh");
+		    	
+		    	if (prerun.exists() == false) {
+		    		try {
+						FileWriter w = new FileWriter(prerun.getAbsolutePath(), false);
+						w.write("#! /usr/bin/sh\n# PreRun script for project " + prjName + "\n\n");
+						w.write("# Place your code below this line.\n");
+						w.write("# Keep your code above this line.\n\n");
+						w.write("exit 0\n\n");
+						
+						w.write("# Function to copy to input directory (jpackage --input).  Usage: CP filepath\n");
+						w.write("CP()\n{\n");
+						w.write("cp $1 " + jg.currPrj.getString("Application Image Options", jg.platform + " Input") + " >/dev/null 2>&1\n");
+						w.write("if [ $? -ne 0 ]; then\n");
+						w.write("    echo PreRun: Copy of $1 failed.\n");
+						w.write("    exit 1\n");
+						w.write("fi\n}\n");
+						w.close();
+						prerun.setExecutable(true);
+					} catch (IOException e) {
+						e.printStackTrace();
+					}
+		    	}
 	    	}
 	    	
 	    	File postrun = new File(System.getProperty("user.home") + File.separator + "JpGui" + File.separator +
@@ -1284,28 +1287,31 @@ public class JpGuiController implements Initializable, RefreshScene {
 				}
 	    	}
 	    	
-	    	postrun = new File(System.getProperty("user.home") + File.separator + "JpGui" + File.separator +
-					"projects" + File.separator + prjName + File.separator + os + "_postrun.sh");
-	    	
-	    	if (postrun.exists() == false) {
-	    		try {
-					FileWriter w = new FileWriter(postrun.getAbsolutePath(), false);
-					w.write("#! /usr/bin/sh\n# PostRun script for project " + prjName + "\n\n");
-					w.write("# Place your code below this line.\n");
-					w.write("# Keep your code above this line.\n\n");
-					w.write("exit 0\n\n");
-					
-					w.write("# Function to delete files.  Usage: REMOVE filepath\n");
-					w.write("REMOVE()\n{\n");
-					w.write("rm -rf $1 >/dev/null 2>&1\n");
-					w.write("if [ $? -ne 0 ]; then\n");
-					w.write("    echo PostRun: Delete of %1 failed.\n");
-					w.write("    exit 1\n");
-					w.write("fi\n}\n");
-					w.close();
-				} catch (IOException e) {
-					e.printStackTrace();
-				}
+	    	if (isWin == false) {
+		    	postrun = new File(System.getProperty("user.home") + File.separator + "JpGui" + File.separator +
+						"projects" + File.separator + prjName + File.separator + os + "_postrun.sh");
+		    	
+		    	if (postrun.exists() == false) {
+		    		try {
+						FileWriter w = new FileWriter(postrun.getAbsolutePath(), false);
+						w.write("#! /usr/bin/sh\n# PostRun script for project " + prjName + "\n\n");
+						w.write("# Place your code below this line.\n");
+						w.write("# Keep your code above this line.\n\n");
+						w.write("exit 0\n\n");
+						
+						w.write("# Function to delete files.  Usage: REMOVE filepath\n");
+						w.write("REMOVE()\n{\n");
+						w.write("rm -rf $1 >/dev/null 2>&1\n");
+						w.write("if [ $? -ne 0 ]; then\n");
+						w.write("    echo PostRun: Delete of %1 failed.\n");
+						w.write("    exit 1\n");
+						w.write("fi\n}\n");
+						w.close();
+						postrun.setExecutable(true);
+					} catch (IOException e) {
+						e.printStackTrace();
+					}
+		    	}
 	    	}
 			
 			if (isWin == true)
@@ -1314,6 +1320,8 @@ public class JpGuiController implements Initializable, RefreshScene {
 				jg.addStatus("Shell script created.");
 			
 			fw.close();
+			if (isWin == false)
+				f.setExecutable(true);
 		} catch (IOException e) {
 			e.printStackTrace();
 		}
